Added authorization for certain methods

+++ care_api/src/group.py
@@ -10,6 +10,8 @@ class Group:
 
 
     def add_group(self, group_id):
+        if not session['is_admin']:
+            abort(403)
         request_data = request.json
         group_doc =  {"group_id":int(group_id),
             "name":request_data['name'],
@@ -20,10 +22,14 @@ class Group:
         return group_doc
 
     def delete_group(self,group_id):
+        if not session['is_admin']:
+            abort(403)
         self.db_handler.delete('group', {"group_id":int(group_id)})
 
 
     def update_group(self, group_id):
+        if not session['is_admin']:
+            abort(403)
         request_data = request.json
         new_val = {"$set": {'name': request_data['name'],'description':request_data['description']}}
         self.db_handler.update('group',{"group_id":int(group_id)}, new_val)
@@ -43,6 +49,8 @@ class Group:
         return recommend_doc
 
     def recommendation_action(self, recommend_id):
+        if not session['is_admin']:
+            abort(403)
         recommendation = self.db_handler.find('recommendation', {'recommendation_id': int(recommend_id)})
         if recommendation is not None:
             self.db_handler.delete('recommendation', {'recommendation_id': int(recommend_id)})
